requirements.txt
streamlit
streamlit-drawable-canvas
opencv-python-headless
torch
torchvision
Pillow
numpy
matplotlib